+++ alchemist-loading/src/test/kotlin/it/unibo/alchemist/test/TestVariablesOverrider.kt
@@ -11,8 +11,8 @@ package it.unibo.alchemist.test
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
+import it.unibo.alchemist.boundary.loader.VariablesOverrider
 import it.unibo.alchemist.boundary.modelproviders.YamlProvider
-import it.unibo.alchemist.model.util.VariablesOverrider
 import org.kaikikm.threadresloader.ResourceLoader
 
 class TestVariablesOverrider : StringSpec({
